+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html
 
 - Non-backward compatible changes:
     - Hexital and Indicator 'append' will now only append to Candles and not sort placement.
+    - Renamed Indicator 'as_list' to 'readings'
 
 - Reworked and optimised 'append' and added 'insert' and 'prepend'
 - Added IndicatorCollection as a way to better reference Indicator's